+++ mslk/gemm/flydsl/_kernels/preshuffle_gemm.py
@@ -20,8 +20,8 @@ from flydsl.expr.typing import T
 from flydsl.runtime.device import get_rocm_arch as get_hip_arch
 from flydsl.utils.smem_allocator import SmemAllocator, SmemPtr
 
-from .mfma_epilogues import mfma_epilog
-from .mfma_preshuffle_pipeline import (
+from mslk.flydsl.kernels.mma.mfma_epilogues import mfma_epilog
+from mslk.flydsl.kernels.mma.mfma_preshuffle_pipeline import (
     _buffer_load_vec,
     buffer_copy_gmem16_dwordx4,
     load_b_pack_k32,
